excel_menu_gui/requirements.txt
PySide6>=6.4.0
openpyxl>=3.1.0
xlrd==1.2.0
xlwings
python-pptx>=0.6.21
PyInstaller>=5.0
Pillow>=9.0.0
